requirements.txt
langchain-community==0.3.22
numpy==1.26.4
openai==0.28
scikit-learn==1.6.1
nltk==3.9.1
sastrawi==1.0.1
gensim==4.3.3
numba==0.61.2
PyPDF2==3.0.1
pdfminer.six==20221105
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
werkzeug==3.0.2
pypdf==5.4.0
pymupdf==1.25.5

# New dependencies for hybrid model
rank-bm25==0.2.2
scipy==1.13.0
pandas==2.2.2
joblib==1.4.2
anyio==4.3.0
python-dateutil==2.9.0.post0